Move resources up to Europeana:: module
