--- lib/europeana/api/record/hierarchy/ancestor_self_siblings.rb
@@ -1,12 +1,0 @@
-module Europeana
-  module API
-    class Record
-      class Hierarchy
-        ##
-        # Retrieve record full(ish) hierarchy data over the Europeana API
-        class AncestorSelfSiblings < Base
-        end
-      end
-    end
-  end
-end
