--- lib/europeana/api/record/hierarchy/children.rb
@@ -1,12 +1,0 @@
-module Europeana
-  module API
-    class Record
-      class Hierarchy
-        ##
-        # Retrieve record children hierarchy data over the Europeana API
-        class Children < Base
-        end
-      end
-    end
-  end
-end
