--- lib/europeana/api/record/hierarchy/following_siblings.rb
@@ -1,12 +1,0 @@
-module Europeana
-  module API
-    class Record
-      class Hierarchy
-        ##
-        # Retrieve record following siblings hierarchy data over the Europeana API
-        class FollowingSiblings < Base
-        end
-      end
-    end
-  end
-end
