--- lib/europeana/api/record/hierarchy/parent.rb
@@ -1,12 +1,0 @@
-module Europeana
-  module API
-    class Record
-      class Hierarchy
-        ##
-        # Retrieve record parent hierarchy data over the Europeana API
-        class Parent < Base
-        end
-      end
-    end
-  end
-end
