--- lib/europeana/api/record/hierarchy/self.rb
@@ -1,12 +1,0 @@
-module Europeana
-  module API
-    class Record
-      class Hierarchy
-        ##
-        # Retrieve record self hierarchy data over the Europeana API
-        class Self < Base
-        end
-      end
-    end
-  end
-end
