+++ lib/europeana/record/hierarchy/ancestor_self_siblings.rb
@@ -1,0 +1,10 @@
+module Europeana
+  class Record
+    class Hierarchy
+      ##
+      # Retrieve record full(ish) hierarchy data over the Europeana API
+      class AncestorSelfSiblings < Base
+      end
+    end
+  end
+end
