+++ lib/europeana/record/hierarchy/children.rb
@@ -1,0 +1,10 @@
+module Europeana
+  class Record
+    class Hierarchy
+      ##
+      # Retrieve record children hierarchy data over the Europeana API
+      class Children < Base
+      end
+    end
+  end
+end
