+++ lib/europeana/record/hierarchy/following_siblings.rb
@@ -1,0 +1,10 @@
+module Europeana
+  class Record
+    class Hierarchy
+      ##
+      # Retrieve record following siblings hierarchy data over the Europeana API
+      class FollowingSiblings < Base
+      end
+    end
+  end
+end
