+++ lib/europeana/record/hierarchy/parent.rb
@@ -1,0 +1,10 @@
+module Europeana
+  class Record
+    class Hierarchy
+      ##
+      # Retrieve record parent hierarchy data over the Europeana API
+      class Parent < Base
+      end
+    end
+  end
+end
