+++ lib/europeana/record/hierarchy/self.rb
@@ -1,0 +1,10 @@
+module Europeana
+  class Record
+    class Hierarchy
+      ##
+      # Retrieve record self hierarchy data over the Europeana API
+      class Self < Base
+      end
+    end
+  end
+end
